ADM: Remove LE Audio hearing aid variables
Local variable mLeHearingAidActiveDevice is prone
to cause errors and is not needed to track active HA.
List of connected and active hearing aids and LE Audio
devices is sufficient to implement all logic.

When we need to determine if HA is LE or ASHA, we can
check if HAP is supported by the device.

This helps with future maintenance and debugging.

Bug: 441143670
Bug: 450220555
Test: atest ActiveDeviceManagerTest
Flag: com.android.bluetooth.flags.adm_remove_hap_variables
Change-Id: I0be0df90073011bfb1b88ddb152578dd66337a48