Add `DoInThreadSynchronously` with deadlock detection to `MessageLoopThread`.
This change introduces a new method, `DoInThreadSynchronously`, which allows posting a task to a `MessageLoopThread` and blocking the caller until the task completes. To prevent deadlocks, a mechanism for detecting circular dependencies between threads is implemented. If a circular dependency is detected (e.g., Thread A is waiting on Thread B, and Thread B tries to wait on Thread A), the task is executed immediately on the caller thread instead of being posted, thus avoiding a deadlock. Static maps are used to track waiting threads, protected by a mutex. Unit tests are added to cover successful synchronous execution and the deadlock detection scenario.

Bug: 447038904
Bug: 418119118
Test: mmm packages/modules/Bluetooth
Test: m com.google.android.bt
Flag: com.android.bluetooth.flags.replace_message_loop_thread_with_gd_handler

Change-Id: I509e81c87829259740ad31a8dcf0c8d8f68199f5