btm_iso: Improve canceling CIS
With this change, btm_iso hides from the btm_iso module user the
complexity of canceling CIS.
When Host cancel CIS the HCI Disconnect command is used and then
controller will send 3 events.

Command Status (always as a first one)
Disconnect Completed
Established Complete Event with a reason HOST CANCELED.

However spec does not say what is the order of last two events.
Currently BTM_ISO sends last two events to the user (leaudio in our
case) and leaudio need to decide how to handle it and nowadays it just
acts on Disconnect Complete event which is almost good.
However with this approach, when order is as presented above, some
controllers might have a problem with following HCI Remove CIG command
as CIS is not yet fully canceled when Disconnect Complete Event is sent.

With this CL, BTM ISO will not send to the user Established Complete
Event with a reason HOST CANCELED, but instead BTM_ISO will track if
both events has been received and if that happens, it will always send
to the user onCisEvent(kIsoEventCisDisconnected).
That means, whenever LeAudio code receives that CIS is disconnected, it
is sure that controller is ready to accept HCI Remove CIG command.

Bug: 447298604
Bug: 450861974
Test: atest bluetooth net_test_btm_iso
Flag: com.android.bluetooth.flags.btm_iso_improve_canceling_iso

Change-Id: I25c16afcc6cd1e60cb53f854dd4e885f27b89015