Prevent NullPointerException in HfpClientConnection
A race condition bug in the constructor could cause a crash. If a call
was terminated right during its creation, we'd end up trying to
initialize a null object. This change adds a simple check to make sure
the connection is still open before proceeding, which stops the crash.

Fix: 451369088
Test: atest HfpClientConnectionTest
Flag: EXEMPT BUGFIX
Change-Id: Id27e8dc9ad9686ce2b6ce39c896694a3ee13fec5