CompanionManager stop using framework metadata callback
The Bluetooth internal code shouldn't use complexe callback pattern to
listen to it's internal change.
Instead we can just intercept the setCustomMetadata call.

Note that this is still a very awful usage of metadata. We should have
exposed a proper API to set the companion device instead of relying on
obfuscated metadata.

Bug: 430166215
Flag: EXEMPT PURE_REFACTOR
Test: atest BluetoothJavaUnitTests
Test: manual migration
Change-Id: I6a6a69640b36a382f081476365f2709f27197aac