Ensure LE scan filter list is never null
This change simplifies the LE scan stack by converting any null filter
list into an empty one at the only API entry point, preventing any
potential `NullPointerException`s, removal of redundant null checks, and
a simpler Kotlin API.

The changes to `ScanClient` are for the reasons of:
Allowing @JvmOverloads to generate multiple constructors with defaulting
to empty scan settings and empty filter list. The `appUid` field is
moved to the first location of the constructor for this to happen.

This simplifies the creation of `ScanClient` both in production code
and in unit tests to avoid passing ScanSettings and List<ScanFilter>
if they are okay with the default values.

Bug: 394958283
Test: atest com.android.bluetooth.le_scan
Flag: EXEMPT PURE_REFACTOR
Change-Id: I6dd5e37a5e94f0a4b2cd528a5f89a4c3eb5df6d8