Release `api_mutex_` while waiting for handler to become idle in `ShutDown`.
The `future.wait_for()` call can block for up to `kHandlerStopTimeout`. Releasing the `api_mutex_` during this wait prevents potential deadlocks or unnecessary blocking of other operations that might require the mutex.
If a task is running on this thread, and from this task another task needs to be posted again it will create deadlock because `api_mutex_` is not available, this will cause future to timeout and assert.

Bug: 409600236
Bug: 418119118
Test: m com.google.android.bt
Flag: com.android.bluetooth.flags.replace_message_loop_thread_with_gd_handler
Change-Id: I17cb14fb6060c851772ed2dc1e8e5be5c2b3729b