btif_hh: instantiate incoming_connection_timer lazily
incoming_connection_timer is initiated on init, and is freed on
cleanup but only if the status is not already DISABLED. However, since
init doesn't immediately assign the status to non-DISABLED, it is
possible for incoming_connection_timer to be initiated but not freed,
leading to memory leak.

This CL instantiate incoming_connection_timer only when it's needed,
and refactor the usage into start_timer and cancel_timer methods.

Bug: 441566940
Test: atest net_test_btif_hh
Flag: EXEMPT BUGFIX
Change-Id: Ic5cccbdac0048ff6013634557adfd2b18443c884