Replace base::Callback and base::Closure with base::Repeating* classes (6)
base::Callback and base::Closure are deprecated and should be replaced.
Replacing with base::Repeating* classes should be a safe no-op operation.
The callbacks are stored in the Reactor::Reactable class and are
potentially called multiple times, so they have to be repeating callbacks.
Also add a few std::move() calls for a slightly better performance and
replace a common::OnceClosure with base::OnceClosure for consistency.

Bug: 303363630
Test: atest packages/modules/Bluetooth
Flag: EXEMPT PURE_REFACTOR

Change-Id: I4159e13591418e070ea9c89af1c3773eed0af498