leaudio: Fix clearing Voip state in context type manager
VoipState in context type manager is set internally when there are
contexts for voice call or ringtone but there is no TBS Signal saying that
this is a regular call.
VoipState is updated on every context type change but it was forgoten to
be updated when context types are cleared.

This could lead to improper state of audio_context_type_manager

Bug: 432138731
Flag: EXEMPT BUGFIX
Test: atest bluetooth_le_audio_client_test
Change-Id: I1b101a633579c92f8b4d3be7189cb03dc33e6b1c