Replace base::Callback with base::RepeatingCallback (12)
base::Callback is deprecated and should be replaced.
Replacing with base::Repeating* classes should be a safe no-op operation.
enqueue_command in LeAddressManager is stored as a class member and
is probably called multiple times, so it cannot be a OnceCallback.
Remove some base:: -> common:: renaming.
Add some std::move() for a slightly better performance.

Bug: 303363630
Test: atest packages/modules/Bluetooth
Flag: EXEMPT PURE_REFACTOR

Change-Id: I9825968b1e66fb1ed6f8ec6849808259c5ee7264