Refactor GetAudioSetConfigurationRequirements logic
Refactor the `GetAudioSetConfiguraitionRequirements` function to improve logic flow and logging.

The logic is reordered to first check for audio direction support before checking for the device's audio locations. This is more efficient as it avoids unnecessary checks for directions that are not supported in the current audio context. The redundant check for audio location has also been removed.

Logging is improved by:
- Adding a new debug log for the `context_type` at the beginning of the function.
- Changing the log level from `warn` to `info` for the "pushes requirement" message

Bug: 446684779
Test: m com.google.android.bt
Flag: EXEMPT PURE_REFACTOR
Change-Id: Ie785dc6de2084e67ff2489adbb2cd9054a02cf29