asrc: use pool with burst buffer rotation
Before this change, the first block in the pool (of size 3) is
initialized to be silent, and burst buffers point to that first block
for every burst block.

However, upon first run, the first block of the pool is immediately
corrupted (filled by the resampled chunks), and until the burst timeout
occurs, the result (`out`) is assigned of that first block on every
iteration.

Assuming the effect we want is to start with `num_burst_buffers` worth
of silent blocks without changing the number of blocks sent, and then
burst all the excess blocks at once on `burst_delay_ms`, in this CL we
ensure the rotation occurs as expected.

Bug: 444568207
Bug: 453508454
Test: m -j com.google.android.bt
Test: play LEA SW and make sure no glitch in beginning
Flag: com.android.bluetooth.flags.lea_asrc_rotate_burst_buffer
Change-Id: I279c9598d307c89c08461194ce74762c553689c8