Skip waiting for UUIDs for LE non-audio devices.
When the bond state changed to bonded, if we don't have the UUIDs for
the remote device we wait some time before broadcasting the bonded
intent so that we have a chance to receive them. With non-audio LE
devices, there is no UUIDs to be received so we can skip waiting immediately.

Bug: 285243669
Bug: 446624664
Flag: com.android.bluetooth.flags.immediate_sdp_results_le
Test: atest BluetoothJavaUnitTests
Change-Id: Ib31b52e96e13676b1bce63bc795d8ab39bb8c5a8