Extract ContextMap.App to its own ContextApp.kt
This is cleanup in preparation for `Flag.gattThread()`. Following the
structure and learnings from `le_scan` with `ScannerMap` and
`ScannerApp`, we need to handle the `uid` correctly when hoping threads.
Currently, there a lot of nested `Binder.getCallingUid` nested within
the classes. This refactor will make it easier to centralize the logic
and inject the uid correctly.

Also remove unused `getAllAppsCallbackId`.

Bug: 453872412
Test: atest com.android.bluetooth.gatt
Flag: EXEMPT PURE_REFACTOR
Change-Id: I588d9149584462826240eb5027d8f587da71aa95