SystemServer: fix some more robolectric tests
Due to jarjaring, it was crashing when looking for some flags.
Due to non-jarjaring, it couldn't use the framework utiliy to send the
message and handle the callback.

I could have move the code to common/... but it require to still be java
for framework compatibility... Long story short, we will have some
duplicate code and we will be happy with that

Bug: 450375470
Flag: EXEMPT TEST_ONLY
Test: atest ServiceBluetoothRoboTests
Change-Id: I3d4739fb45099ee1d28438ec852aa5d7ac696348