Convert ContextMap to Kotlin
Super unoptimized, just pure Java to Kotlin translation. To keep it
minimal friction and as exact behavior as possible with the existing
Java class

This is cleanup in preparation for `Flag.gattThread()`. Following the
structure and learnings from `le_scan` with `ScannerMap` and
`ScannerApp`, we need to handle the `uid` correctly when hoping threads.
Currently, there a lot of nested `Binder.getCallingUid` nested within
the classes. This refactor will make it easier to centralize the logic
and inject the uid correctly.

Bug: 453872412
Test: atest com.android.bluetooth.gatt
Flag: EXEMPT PURE_REFACTOR
Change-Id: Ie081a8afd487f345eb224ae8de91e8608aa063f5