Pass over `uid` from GattService to AdvertiserMap
When a task is posted via `doOnGattThread(() -> foo()`, if we do
`Binder.getCallingUid()`, that returns our own app uid because now we
are on the Gatt thread. So we have to use the uid from an
`AttributionSource`.

This unfortunately led to the requirement of AttributionSource which
had a cascading affect across multiple classes but it had to be done.

Also a small followup of ag/36503115

Bug: 449681465
Bug: 453872412
Test: atest com.android.bluetooth.gatt
Flag: com.android.bluetooth.flags.gatt_thread
Change-Id: Ic892f3d8dd342f33f715393e25b7239b562143f5