LeAudio: Introduce ISO proxy for the Peripheral role
Adds `IsoProxy`, a helper class to manage interactions with the
IsoManager for the LE Audio Peripheral role.

This proxy encapsulates Isochronous (ISO) channel state tracking and
simplifies the API for sending and receiving data. This removes the
need for the upcoming AseManager to manage this logic directly.

Bug: 428719427
Bug: 424275371
Test: atest bluetooth_le_audio_server_test
Flag: com.android.bluetooth.flags.leaudio_peripheral_feature
Change-Id: Ie4ee541d59a0bbb9bbad495b22340e83f5ef3a4f