Pending HCI Delete Stored Link Key command must not fail pairing
HCI Delete Stored Link Key command is sent to the controller to get the controller to delete any stored link keys and force it to ask for the link key from the host if authentication is requested.
There is no need to wait for its response or serialize these commands.

Test: mmm packages/modules/Bluetooth
Flag: EXEMPT BUGFIX
Bug: 393200103
Change-Id: I28e394f9f481f5da09d77612f8bb848a9acbff44