Post error callback on scan registration failure
When the `scanRegisterAndStart` flag is enabled, a scanner
registration failure (any status other than GATT_SUCCESS)
did not notify the client. This change adds the missing call to
`postCallbackError` with `SCAN_FAILED_APPLICATION_REGISTRATION_FAILED`
to ensure the callback is invoked on failure.

Additionally, the `mScannerId` is not being considred now because it
will not be set on failure. Whereas previously, it would have a 2
second timeout and get set regardless of success / failure.

This failure was detected during flag advancement: ag/36577257

Bug: 455057044
Test: atest com.android.bluetooth.le_scan
Flag: com.android.bluetooth.flags.scan_register_and_start
Change-Id: Id3c233a60d6304662b0f7c32ae3f1c00f69a6597