Update resource names to be agnostic of A2DP/AVRCP
Bluetooth Media is soon going to be started agnostic of A2DP/AVRCP, in
support of LE Audio.

Bluetooth Media has an app name and a queue name. The old string
resource names were based on A2DP Sink.

This change cleans up the naming to be agnostic of a given profile.

Flag: EXEMPT PURE_REFACTOR
Bug: 450371968
Test: m -j
Test: atest com.android.bluetooth.avrcpcontroller
Change-Id: I8ab900787b0561acbc3ccc0d38ff1fd6e692f359