Avoid storing never [later] accessed isQApp boolean in ScanClient
This was made very evident thanks to ag/36650018:
  - Rename private startScan to dispatchStartScan in ScanController

Clearly seeing that this specific `startScan`, never gets called again.
Therefore, the `ScanClient` is set but never accessed outside of it.
Hence, it can be made a local variable and removed from the class.

To double check, that we did not do any recent mistakes, I went back one
year to `20241031` and confirmed the same lack of later access at that
date as well.

Bug: 394958283
Test: atest BluetoothJavaUnitTests
Test: atest com.android.bluetooth.le_scan
Flag: EXEMPT PURE_REFACTOR
Change-Id: Ie8b46ed2dd317b820ddf28f0842f63cafbdda955