Move BluetoothMediaBrowserService from avrcpcontroller to media_audio
BluetoothMediaBrowserService is about to be used by AVRCP Controller and
up and coming MCP Client code. It cannot be specific to AVRCP Controller
any more and needs to be in a more central location for everyone to use.

A refactor of the code is coming after the move.

Flag: EXEMPT PURE_REFACTOR
Bug: 450371968
Test: atest com.android.bluetooth.avrcpcontroller
Test: m -j
Change-Id: Idc0d4098ab7d7d5d0c4d362b3f1076c88c9f5586