Refactor Shuffle and Repeat handling not to use getSession
Problem: AVRCP Controller's shuffle and repeat comes from player
application settings changes. We directly access the media session from
without the AvrcpPlayer abstraction to post updates to the shuffle and
repeat status. This breaks the abstractions. We do not want to expose
getSession() long term. As well, these calls are not protected by a call
to isActive(), which risks us updating shuffle and repeat status for the
session when a given device isn't active. This would let us step on a
different active device. As well, we don't update shuffle and repeat
status on active device change, which means we can get out of sync

Solution: Move BluetoothMediaBrowserService updates of shuffle and
repeat status to the state machine. Use isActive() to protect updates.
Be sure to update shuffle and repeat on active device change. Add new
functions to BluetoothMediaBrowserService to handle shuffle and repeat,
and use those in AVRCP Controller.

Flag: EXEMPT BUGFIX
Bug: 450371968
Test: atest com.android.bluetooth.avrcpcontroller
Change-Id: I055c9f32a29d2a97119bfb56d442d5c86df8a419