Remove mMediaQueue and use a local copy instead
The member variable copy of the queue isn't necessary. We just need to
provide a copy of the list to the Media Framework so no one edits the
order of the list in AVRCP. As well, the debug print of it was verbose
and unneeded.

With mMediaQueue gone, clearNowPlayingQueue can be removed and simple
calls to mSession.setQueue(null), setNowPlayingQueue(null), or
setNowPlayingQueue(emptyList()) can all be used instead. Note an empty
list maps to null automatically in the function call.

Flag: EXEMPT PURE_REFACTOR
Bug: 450371968
Test: atest com.android.bluetooth.avrcpcontroller
Change-Id: I3704282bdbda8f049f725461d17566838784bfc2