Introduce ActionOnDeathRecipient
I noticed this pattern of an almost-data class in `le_scan` and `gatt`.
Unfortunately, they cannot be a data-class due to the custom actions but
then I realized, they dont need to be a unique subclass at all.

Unfortunately due to the Java usage, we needed a custom constructor
which took a Runnable and while it is still not as clean as concise
as the Kotlin implementation, I think it vastly cleansup boilerplate
code and a convenient, standardized logging.

Fix: 455679694
Test: atest BluetoothJavaUnitTests
Flag: EXEMPT PURE_REFACTOR
Change-Id: I73493004734229be58ecd146683893549396ebf7