Replace base::Callback with modern alternatives (19)
base::Callback is deprecated and should be replaced.
Refactor the code around MsftAdvMonitor*Callback callbacks
to never store them.
This makes it obvious that they are never called more than
once and can be replaced with OnceCallbacks.
The callback was not stored in all the “if” branches, which is
probably a bug. This is now fixed.

Bug: 303363630
Test: atest packages/modules/Bluetooth
Flag: EXEMPT PURE_REFACTOR

Change-Id: I242ffb5132015777456ea367efc672e22158fc32