Rename `checkServiceAvailable` to `checkProfileAvailable`
They have not been Android services for quiiite a long time now.
Also move to `Util.kt` and make it an extension for convenient Kotlin
usage.

Ideally, this function should not have to check for `null` input. But
that is a cleanup for when `android/app` is 100% Kotlin.

Bug: 455679694
Test: atest com.android.bluetooth.btservice
Flag: EXEMPT PURE_REFACTOR
Change-Id: I30ddeff25604bcafca5634f46ce5388d0f7aab8d