Optimize ScanController (doOnScanThread) calls from BassClientService
Adds a new method `ScanController.isCallerOnScanThread()` to check
if the current thread is the scan handler's thread. If a callback in
BassClientService is already on the scan thread, it now calls these
methods directly to prevent a crash:
  java.lang.IllegalStateException: Must NOT be on scan thread

This effectively reverts the changes in ag/36265587 as well.

Fix: 457779248
Bug: 397863857
Test: atest com.android.bluetooth.bass_client
Flag: com.android.bluetooth.flags.scan_controller_thread
Change-Id: I05bb13d1cc9640bc63a06ff0054107589666738e