TableBuilder calls `toString()` on `Any` object
This allows omitting the call to `.toString()` when passing over
objects. This will allow both normal objects to print properly and for
Kotlin value classess to not have to call `.toString()`.

Also rename `c` amd `w` as even forgetting the code for a couple of
weeks lead to confusion :) I can't imagine how it'll be looking back at
it after a year.

Bug: 446315696
Test: atest BluetoothJavaUnitTests
Test: manual via observing dumpsys output
Flag: EXEMPT PURE_REFACTOR
Change-Id: I2e2aa3eb207e4bcb555c2b4f20fe9a1a1757c244