hid host: Cancel GATT connection when policy is set to forbidden
When we set the connection policy to forbidden, we correctly
disconnect the device and prevent future profile connection. However,
if at that time we already put the device into the accept list, the
device won't be taken out of the accept list, so it can still connect,
albeit only for a very brief moment before it is disconnected because
of no accepting profile.

This CL also cancel the GATT connection, so the device will be taken
out of the accept list.

Bug: 458253534
Bug: 459303232
Test: m com.google.android.bt
Test: Manually, verify HID device is removed from accept list when
      policy is set to forbidden.
Flag: com.android.bluetooth.flags.hogp_cancel_gatt_if_policy_forbidden
Change-Id: Ia55cb5019a7517704cf04d2a616aba37f2ef19c9