Finalize most fields of `ScanClient`
Ideally, `AppScanStats` should be a constructor parameter as well but
due to `ScanClient` class sometimes being used as a `scannerId` wrapper,
an `AppScanStats` object is hard to reach in some places.

Unfortunately, `ScanControllerTest` is configured to be "mutating"
fields and will have to be cleaned up before the last 3 fields can be
made `val`.

Bug: 394958283
Test: atest com.android.bluetooth.le_scan
Flag: EXEMPT PURE_REFACTOR
Change-Id: Ie570a4213cf9d84f5bccd5845b019692b8a8e36f