Remove @JvmOverloads from AppScanStats/ScannerMap
This is in prep for the next CL: Convert `ScanControllerTest`. So,
apparently, when a function is annotated `@@JvmOverloads`, Kotlin is smart enough to choose the
variables (via named parameters) and uses a single function to do but Java only defaults to the one
that matches the signature, resulting in the generation of multiple functions. However, attempting
to `.verify` this in a Kotlin unit test, we are unable to do so because from Kotlin, we call the
only available function that has all the parameters. And so the `.verify` fails.

Bug: 458821587
Test: atest com.android.bluetooth.le_scan
Flag: EXEMPT PURE_REFACTOR
Change-Id: I0153e0a4023d92f06bcafcd08ceeec2a60e429cc