Introduce inline Kotlin `Util.kt#callbackToApp`
Adds a Kotlin-idiomatic utility function using `inline`. This replaces
the Java version (available only for Kotlin files), avoiding the runtime
allocation of anonymous objects for lambdas and reducing overhead!

A detailed JVM decompiled bytecode analysis can be found at
https://b.corp.google.com/issues/455679694#comment18

`GattServerManager` is the first class to make use of this! 🎉
For now, it is also the only class that can actually make use of this.

Bug: 455679694
Test: atest UtilsTest
Flag: EXEMPT PURE_REFACTOR
Change-Id: I689a6cef279e164d877f20a22b5f767113438a7d