Add a new `ScanSettings#toBuilder` extension to fix lost scan filters
If a previous scan request came in with `setReportDelay` of 3000 and RSSI 50, they would get the `setReportDelay` set to 5000 and lose the RSSI filter search. This fixes the incorrect behavior.

Add a Kotlin extension function to provide a consistent way of copying
previous various `ScanSettings` fields instead of manually copying
(and missing a few) fields when updating a single setting.

Fix: 458164203
Test: atest com.android.bluetooth.le_scan
Flag: EXEMPT BUGFIX
Change-Id: I4bd2836e3da894d81867fd2bbc62fe878ccab5c2