Log GATT Offload session state changes.
This change introduces metrics logging for GATT Offload sessions. It
adds a new function `LogGattOffloadSessionStateChanged` to
`os_metrics.h` and implements it in `metrics.cc`. The `gatt_offload.cc`
module is updated to call this logging function when a GATT Offload
session is started (`add_hal_session`) and when it is stopped or fails
(`remove_hal_session`).

go/gatt_offload_metrics

Bug: 456228282
Test: m com.google.android.bt
Test: atest net_test_stack_gatt_offload
Flag: EXEMPT METRICS
Change-Id: I315f84303691aae4fdaba31f746c14788334b47f