[Scan Multiplexing] Migrate and edit scan state keeping functionality to GD
Scan state keeping for observe and inquiry is currently done in BTM.
Migrate and edit this functionality to GD for Scan Multiplexing feature.
Functions are changed to keep states of 3 types of scan: Java scan, CSIS
scan and Discovery. BTM code is scheduled for delete when the feature is
complete.

Bug: 358741228
Test: m -j;
Flag: EXEMPT PURE_REFACTOR
Change-Id: Id56f8f532260bc327f8915cffdfca0ddb220c6e3