Fix LE Audio Config Selection for Sink Only Requirements
A sink direction only configuration could be incorrectly chosen when the
requirement specified a bidirectional stream. It's because the function
`IsAudioSetConfigurationSupported` just ignore the check when a
configuration is lack of that direction.

This change adds a check to ensure that a sink only configuration will
be rejected for a bidirectional requirement.

Bug: 443136442
Test: m com.google.android.bt
Flag: EXEMPT BUGFIX
Change-Id: I564ddf062d33ea84a2b241bc6e73ea87fe7e6550