Rename `continuePiStartScan` as `dispatchPendingIntentStartScan`
The word "continue" was vague; it may imply resuming a paused scan.
However, scan had never started prior to this. We now prefix with
`dispatch` which shows the true implementation of the function: Which
is informing `ScanManager` to start the scan.

And use `PendingIntent` instead of `PI` because why not.

Similar to ag/36650018

Bug: 394958283
Test: atest com.android.bluetooth.le_scan
Flag: EXEMPT PURE_REFACTOR
Change-Id: Ie1e518871587f0b8f5f0c984873ee6acb80e01e9